refactor: extract parser blocks to models/parser_blocks.py
Pure move of FileParserBlock and PDFParserBlock out of block.py into a
new models/parser_blocks.py, with block.py re-exporting both names so
call sites are unchanged. Prunes parser-only preamble imports from
block.py (pandas/docx/csv/codecs/charset-normalizer, the PDF/sanitize
utils, InvalidFileType, and the MAX_FILE_PARSE_INPUT_TOKENS/
MAX_PDF_OCR_PAGES/PDF_OCR_PAGE_CONCURRENCY constants); keeps FileType
re-exported for existing importers.
